djangochallenge/requirements.txt
asgiref==3.2.10
certifi==2020.11.8
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
model-bakery==1.2.1
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
